fix: tons of tx-related improvements. fixes #728 and #729
