+++ packages/app/src/components/popup/tx-item.tsx
@@ -51,7 +51,7 @@ export const TxItem: React.FC<TxItemProps> = ({ transaction }) => {
       </Text>
       <Flex dir="row">
         <Box flexGrow={1}>
-          <Text display="block" fontWeight="400" color="ink.1000">
+          <Text display="block" fontWeight="400" fontSize={2} color="ink.1000">
             {getTxLabel(transaction)}
           </Text>
           <Text fontSize={0} color="ink.400">
